mqtt_to_kafka/requirements.txt
# mqtt_to_kafka/requirements.txt
paho-mqtt>=2.0
kafka-python